update - solv setup
